+++ libraries/install_libyaml.php
@@ -1,18 +1,27 @@
 <?php
 
-
+/**
+ * ██╗     ██╗██████╗ ██╗   ██╗ █████╗ ███╗   ███╗██╗     
+ * ██║     ██║██╔══██╗╚██╗ ██╔╝██╔══██╗████╗ ████║██║     
+ * ██║     ██║██████╔╝ ╚████╔╝ ███████║██╔████╔██║██║     
+ * ██║     ██║██╔══██╗  ╚██╔╝  ██╔══██║██║╚██╔╝██║██║     
+ * ███████╗██║██████╔╝   ██║   ██║  ██║██║ ╚═╝ ██║███████╗
+ * ╚══════╝╚═╝╚═════╝    ╚═╝   ╚═╝  ╚═╝╚═╝     ╚═╝╚══════╝
+ */
 
 
 $path = ARCH_PATH . $lib->name . '-' . $lib->version . '\\';
 $yamllog = LOG . 'libyaml.log';
 
 
+// Verify if libyaml is installed
 if (is_dir($path) && is_file($path . 'build\lib\libyaml.lib') && is_file(DEPS_PATH . 'lib\libyaml.lib')) {
     draw_status($lib->name . '-' . $lib->version, "installed", Green);
     return;
 }
 
 
+// Download and unzip libyaml
 $tmpfile = TMP.pathinfo($lib->download_url, PATHINFO_BASENAME);
 if(!download_file($lib->download_url, $tmpfile, pathinfo($tmpfile, PATHINFO_BASENAME))) exit_error();
 if(!$firstdir = zip_first_dir($tmpfile)) exit_error("Invalid zip archive");
@@ -20,68 +29,32 @@ if(!unzip($tmpfile, ARCH_PATH)) exit_error();
 if(!rename_wait(ARCH_PATH . $firstdir, $path)) exit_error("Can't rename library path");
 
 
-
+// Compile libyaml
 $label = "Compile " . $lib->name . '-' . $lib->version;
 draw_line($label, "running", Yellow);
-
 $bat = '@echo off'.RN;
 $bat .= 'cd ' . escapeshellarg($path).RN;
 $bat .= 'cmake .'.RN;
 $bat .= 'cmake --build . --config Release --clean-first'.RN;
-
-$batfile = TMP . 'build_nghttp2.bat';
+$batfile = TMP . 'build_libyaml.bat';
 file_put_contents($batfile, $bat);
-$ret = shell_exec_vs16($batfile, true);
+$ret = shell_exec_vs16($batfile);
 file_put_contents($yamllog, $ret);
 
 
-if(!is_file($path . 'Release\yaml.lib')) {
-    draw_status($label, "failed", Red, true);
-} else {
-    draw_status($label, "complete", Green);
-}
+// Verify if the build works
+if(!is_file($path . 'Release\yaml.lib')) draw_status($label, "failed", Red, true, 'SEE: ' . $yamllog);
+else draw_status($label, "complete", Green);
 
 
-
-
+// Install libyaml
 $label = "Install " . $lib->name . '-' . $lib->version;
 draw_line($label, "running", Yellow);
-
 $builddir = $path . 'build\\';
-if(!is_dir($builddir) && !@mkdir($builddir, 0777, true))
-    draw_status($label, "failed", Red, true);
+$files[$path . 'Release\yaml.lib'] = 'lib\libyaml.lib';
+$files[$path . 'include\yaml.h'] = 'include\yaml.h';
+if(!create_build($builddir, $files)) draw_status($label, "failed", Red, true);
+if(!install_deps($builddir)) draw_status($label, "failed", Red, true);
+else draw_status($label, "complete", Green);
 
-$libdir = $builddir . 'lib\\';
-if(!is_dir($libdir) && !@mkdir($libdir, 0777, true))
-    draw_status($label, "failed", Red, true);
-
-$incdir = $builddir . 'include\\';
-if(!is_dir($incdir) && !@mkdir($incdir, 0777, true))
-    draw_status($label, "failed", Red, true);
-
-
-
-if(!@copy($path . 'Release\yaml.lib', $libdir . 'libyaml.lib'))
-    draw_status($label, "failed", Red, true);
-
-if(!@copy($path . 'include\yaml.h', $incdir . 'yaml.h'))
-    draw_status($label, "failed", Red, true);
-
-
-
-
-if(!install_deps($builddir)) {
-    draw_status($label, "failed", Red, true);
-} else {
-    draw_status($label, "complete", Green);
-}
-
-
-
-
-
-
-
-
-
-// die();
+delete_parent_deps($lib->name);