+++ libraries/install_pthread.php
@@ -1,10 +1,20 @@
 <?php
 
+/**
+ * ██████╗ ████████╗██╗  ██╗██████╗ ███████╗ █████╗ ██████╗ 
+ * ██╔══██╗╚══██╔══╝██║  ██║██╔══██╗██╔════╝██╔══██╗██╔══██╗
+ * ██████╔╝   ██║   ███████║██████╔╝█████╗  ███████║██║  ██║
+ * ██╔═══╝    ██║   ██╔══██║██╔══██╗██╔══╝  ██╔══██║██║  ██║
+ * ██║        ██║   ██║  ██║██║  ██║███████╗██║  ██║██████╔╝
+ * ╚═╝        ╚═╝   ╚═╝  ╚═╝╚═╝  ╚═╝╚══════╝╚═╝  ╚═╝╚═════╝ 
+ */
+
+
 $path = ARCH_PATH . $lib->name . '-' . $lib->version . '\\';
 $ptlog = LOG . 'pthread.log';
 
 
-// Verify if pthreads is installed
+// Verify if pthread is installed
 if (is_dir($path) && is_file($path . 'build\lib\pthreadVC3.lib') && is_file(DEPS_PATH . 'lib\pthreadVC3.lib')) {
     draw_status($lib->name . '-' . $lib->version, "installed", Green);
     return;
@@ -22,7 +32,6 @@ if(!rename_wait(ARCH_PATH . $firstdir, $path)) exit_error("Can't rename library 
 // Compile pthdreads
 $label = "Compile " . $lib->name . '-' . $lib->version;
 draw_line($label, "running", Yellow);
-
 $slndir = $path . 'windows\VS2019\\';
 $sln = $slndir . 'pthread.2019.sln';
 if(!is_file($sln)) draw_status($label, "failed", Red, true, "Solution file not found");
@@ -37,62 +46,25 @@ file_put_contents($ptlog, $ret);
 
 // Verify if the build works
 $compdir = $slndir . 'bin\Release-Unicode-64bit-x64\\';
-if(!is_file($compdir . 'pthread_static_lib.lib')) {
-    draw_status($label, "failed", Red, true, 'SEE: ' . $ptlog);
-} else {
-    draw_status($label, "complete", Green);
-}
+if(!is_file($compdir . 'pthread_static_lib.lib')) draw_status($label, "failed", Red, true, 'SEE: ' . $ptlog);
+else draw_status($label, "complete", Green);
 
 
 // Install library
 $label = "Install " . $lib->name . '-' . $lib->version;
 draw_line($label, "running", Yellow);
-
-
 $builddir = $path . 'build\\';
-if(!is_dir($builddir) && !@mkdir($builddir, 0777, true))
-    draw_status($label, "failed", Red, true);
-
-$libdir = $builddir . 'lib\\';
-if(!is_dir($libdir) && !@mkdir($libdir, 0777, true))
-    draw_status($label, "failed", Red, true);
-
-$incdir = $builddir . 'include\\';
-if(!is_dir($incdir) && !@mkdir($incdir, 0777, true))
-    draw_status($label, "failed", Red, true);
-
-
-if(!@copy($compdir . 'pthread_static_lib.lib', $libdir . 'pthreadVC3.lib'))
-    draw_status($label, "failed", Red, true);
-
-if(!@copy($compdir . 'pthread_static_lib.lib', $libdir . 'pthreadVC2.lib'))
-    draw_status($label, "failed", Red, true);
-
-if(!@copy($path . '_ptw32.h', $incdir . '_ptw32.h'))
-    draw_status($label, "failed", Red, true);
-
-if(!@copy($path . 'pthread.h', $incdir . 'pthread.h'))
-    draw_status($label, "failed", Red, true);
-
-if(!@copy($path . 'sched.h', $incdir . 'sched.h'))
-    draw_status($label, "failed", Red, true);
-
-if(!@copy($path . 'semaphore.h', $incdir . 'semaphore.h'))
-    draw_status($label, "failed", Red, true);
-
-
-// PATCH sched.h
-$contents = file_get_contents($incdir . 'sched.h');
+copy($compdir . 'pthread_static_lib.lib', $compdir . 'pthreadVC2.lib');
+$files[$compdir . 'pthread_static_lib.lib'] = 'lib\pthreadVC3.lib';
+$files[$compdir . 'pthreadVC2.lib'] = 'lib\pthreadVC2.lib';
+$files[$path . '_ptw32.h'] = 'include\_ptw32.h';
+$files[$path . 'sched.h'] = 'include\sched.h';
+$files[$path . 'semaphore.h'] = 'include\semaphore.h';
+if(!create_build($builddir, $files)) draw_status($label, "failed", Red, true);
+$contents = file_get_contents($builddir . 'include\sched.h');
 $contents = str_replace('  typedef __int64', '  //typedef __int64', $contents);
-file_put_contents($incdir . 'sched.h', $contents);
+file_put_contents($builddir . 'include\sched.h', $contents);
+if(!install_deps($builddir)) draw_status($label, "failed", Red, true);
+else draw_status($label, "complete", Green);
 
-
-if(!install_deps($builddir)) {
-    draw_status($label, "failed", Red, true);
-} else {
-    draw_status($label, "complete", Green);
-}
-    
 delete_parent_deps($lib->name);
-    
-