+++ php/install_php.php
@@ -1,7 +1,18 @@
 <?php
 
+/**
+ * ██╗███╗   ██╗███████╗████████╗ █████╗ ██╗     ██╗         ██████╗ ██╗  ██╗██████╗ 
+ * ██║████╗  ██║██╔════╝╚══██╔══╝██╔══██╗██║     ██║         ██╔══██╗██║  ██║██╔══██╗
+ * ██║██╔██╗ ██║███████╗   ██║   ███████║██║     ██║         ██████╔╝███████║██████╔╝
+ * ██║██║╚██╗██║╚════██║   ██║   ██╔══██║██║     ██║         ██╔═══╝ ██╔══██║██╔═══╝ 
+ * ██║██║ ╚████║███████║   ██║   ██║  ██║███████╗███████╗    ██║     ██║  ██║██║     
+ * ╚═╝╚═╝  ╚═══╝╚══════╝   ╚═╝   ╚═╝  ╚═╝╚══════╝╚══════╝    ╚═╝     ╚═╝  ╚═╝╚═╝     
+ */
+
+
 $path = ARCH_PATH . 'php-' . $MATRIX->php->version . '-src\\';
 
+// Download and unzip PHP
 if(!is_dir($path)) {
     if(curl_file_exists($MATRIX->php->download_url)) $url = $MATRIX->php->download_url;
     else $url = $MATRIX->php->archive_url;
@@ -10,28 +21,25 @@ if(!is_dir($path)) {
     if(!unzip($tmpfile, ARCH_PATH)) exit_error();
     if(!is_dir($path)) exit_error("Can't find PHP folder");
 }
-
 draw_status("PHP-" . $MATRIX->php->version, 'installed', Green);
 
 
-
+// Apply patches
+if(SPEED_DEV) return;
 $patchdir = DIR.'master\patches\\' . $MATRIX->php->version . '\\';
 if(!is_dir($patchdir)) exit_error("Can't find PHP-' . $MATRIX->php->version . ' patches folder");
-
 $patchexe = DIR . 'sdk\msys2\usr\bin\patch.exe';
 if(!is_file($patchexe)) exit_error("Can't find patch.exe");
 
-if(!SPEED_DEV) {
-    draw_line("Apply patches", 'running', Yellow);
-    
-    $bat = '@echo off'.RN;
-    foreach(glob($patchdir.'*.patch') as $patch) {
-        $bat .= escapeshellarg($patchexe) . ' --force --directory=' . escapeshellarg($path) . ' -p 2 < ' . escapeshellarg($patch).RN;
-    }
+draw_line("Apply patches", 'running', Yellow);
 
-    $batfile = TMP . 'php_patch.bat';
-    file_put_contents($batfile, $bat);
-    shell_exec(escapeshellarg($batfile) . ' 2>&1');
-
-    draw_status("Apply patches", 'complete', Green);
+$bat = '@echo off'.RN;
+foreach(glob($patchdir.'*.patch') as $patch) {
+    $bat .= escapeshellarg($patchexe) . ' --force --directory=' . escapeshellarg($path) . ' -p 2 < ' . escapeshellarg($patch).RN;
 }
+
+$batfile = TMP . 'php_patch.bat';
+file_put_contents($batfile, $bat);
+shell_exec(escapeshellarg($batfile) . ' 2>&1');
+
+draw_status("Apply patches", 'complete', Green);