+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 # Change Log
 
-## yyyy-mm-dd
+## [Unreleased] - yyyy-mm-dd
 
 ### Added
+- haru pdf support
+- dio serial support
+- sync support
+- win32service support
+- win32ps support
 
 ### Changed
+- git clone and pull recursive
 
 ### Fixed
+- draw_status label fix
+- yaml build script name